refactor: changed the directory display
