+++ ciudadano_app/templates/bienvenida.html
@@ -12,7 +12,7 @@
           <span class="font-bold text-xl">Panel de Entidad Municipal</span>
         </div>
         <div class="flex items-center">
-          <a href="{% url 'logout' %}" class="bg-red-600 px-4 py-2 rounded hover:bg-red-500 text-sm font-medium">
+          <a href="{% url 'logout_ciudadano' %}" class="bg-red-600 px-4 py-2 rounded hover:bg-red-500 text-sm font-medium">
             Cerrar Sesión
           </a>
         </div>
