+++ ciudadano_app/urls.py
@@ -3,8 +3,11 @@ from . import views
 
 urlpatterns = [
     # URLs para autenticación de ciudadano
-    path('login/', views.login_ciudadano, name='login_ciudadano'),
-    path('registro/', views.registro_ciudadano, name='registro_ciudadano'),
-    path('logout/', views.logout_ciudadano, name='logout'),
     path('', views.bienvenida_ciudadano, name='bienvenida_ciudadano'),
+    path('registro/', views.registro_ciudadano, name='registro_ciudadano'),
+    path('dashboard/', views.dashboard_ciudadano, name='dashboard_ciudadano'),
+    path('bienvenida/', views.bienvenida_ciudadano, name='bienvenida_ciudadano'),
+    path('registro/', views.registro_ciudadano, name='registro_ciudadano'),
+    path('login/', views.login_ciudadano, name='login_ciudadano'),
+    path('logout/', views.logout_ciudadano, name='logout_ciudadano'),
 ]
