+++ ciudadano_app/views/__init__.py
@@ -1,0 +1,13 @@
+from .dashboard import dashboard_ciudadano
+from .registro import registro_ciudadano
+from .login import login_ciudadano
+from .logout import logout_ciudadano
+from .bienvenida import bienvenida_ciudadano
+
+__all__ = [
+    'dashboard_ciudadano',
+    'registro_ciudadano',
+    'login_ciudadano',
+    'logout_ciudadano',
+    'bienvenida_ciudadano',
+]
